+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - styles for toolbar
+- styles for overallInfo
 
 ## [3.3.0] - 2022-01-06
 ### Added